superscript and links in html
I am so done trying to make sense of branches right now. working in master only at the moment.

I added superscript text to the menu items and made all of the thumbnail images to links, and made sure that the one for the project page I wanted to do actually works and goes to the project page, and the home link on the project page works. woo!