CSS redo
Overwhelmed by my notes and all of the things I don't understand in the boilerplate CSS that I have to scroll to before I see something familiar, I decided to simplify, getting rid of my notes and designing for the smallest range of widths first and then adding rules as it scales up. My detailed notes are still in the master branch, so I can refer to those for breaking points and specifics, etc.